Update index.md
It isn't made clear enough that android returns to you a FCM token when iOS doesn't. I've watched several people in the community have to figure this inconsistency out for themselves when a note could be made here and save all of them time, so let me know if I would need to go about including this information another way.