Rename applinks.json to assetlinks.json
When creating `applinks.json` (as opposed to Google's requirement to call the file `assetlinks.json`, I get the following error when clicking "Test statement" in the Google Statement List Generator and Tester:

> No app deep linking permission found for nl.highfivefoundation.app at app.highfivefoundation.nl.

Renaming to `assetlinks.json` (like Google also recommends) solves this issue. Updated the docs to reflect this change.