fix(Splash): parse colors as RGBA
On iOS, colors are parsed as RGBA. I would argue that this is the way it **should** be, because that is the standard CSS format (as opposed to ARGB). As web developers, when we see a hex color prefixed by "#", we expect it to be a standard CSS color, which means RGBA.

On Android, `Color.parseColor()` is backwards; it expects ARGB. This is especially surprising coming from google. In any case, this PR does two things to make colors consistent across platforms:

- Allows the "#" to be optional on Android (as it is on iOS).
- Parses the color as RGBA on Android.