fix(ios): Use autolayout for CAPWebView and JSValueContainer Update (#5487)
fix(ios): Add `getValue(_:)` method to JSValueContainer. This intends to replace `getAny(_:)`, since we are not ever going to be returning a type that does not conform to `JSValue`.

fix(ios): Update CAPWebView to use autolayout to pin the child webview to it's leading, trailing, top, and bottom constraints. This was causing issues with when CAPWebView is nested deep within a view heirarchy. Refactor to make `webView` and `capacitorBridge` to no longer be optionals through the use of lazy vars.