Capacitor HTTP fetch API support
The Capacitor HTTP plugin currently requires that all request to the browser fetch API is in the form ```fetch(resource: string, options: object)``` in order to proxy request through the native HTTP plugin. This is confusing, as the fetch API supports other formats like ```fetch(resource: Request)``` which is then not proxied, but instead sent through the browser fetch API directly which may leads to CORS issues.