fix(ios): fix boundary value extraction for form-data requests
The changes introduced in #7306 fixed several bugs with multipart/form-data requests on iOS.
However, the extraction of the actual boundary value might fail due to the value
being surrounded by double quotes, which is allowed and happens occasionally.
Additionally, the `Content-Type` header might include other keys such as `charset`.
This change extracts the boundary value regardless of the order of the individual keys within
the `Content-Type` header.